red herring comment deleted
This is the part I was referring to in my email, and when we were talking in Tempe. But now I see at the top of the file it came from X Consortium so I imagine you just left some of these comments. I spent a ton of time trying to figure out a solution to this umask gripe though, and in the end I think umask works exactly how it's intended. And is a great security feature.